requirements.txt
python-dotenv
python-telegram-bot
mysql-connector-python
argparse
discord

